UX: Display confirmation message after subscribing (#2)
Currently, when a user subscribes to the newsletter via the banner at top of the site, the banner disappears as soon as the subscribe button is clicked. This behavior isn't ideal from a UX perspective because there's no confirmation that the user action has done anything.

This commit improves the subscribe UX by displaying a confirmation message after the user clicks the subscribe button. It also moves the 'Manage preferences' link from the initial banner state to the confirmation message.

Internal topic: t/96002/31.

+++ assets/javascripts/discourse/connectors/top-notices/newsletter-banner.js
@@ -8,8 +8,10 @@ import { popupAjaxError } from "discourse/lib/ajax-error";
 export default class NewsletterBanner extends Component {
   @service site;
   @service currentUser;
+
   @tracked disableControls = false;
   @tracked dismissed = false;
+  @tracked subscribed = false;
 
   get showBanner() {
     return (
@@ -23,26 +25,26 @@ export default class NewsletterBanner extends Component {
   async subscribe() {
     this.disableControls = true;
 
-    let succeeded = true;
     try {
       await ajax("/newsletter-integration/subscriptions", { type: "POST" });
+      this.subscribed = true;
     } catch (e) {
-      succeeded = false;
       popupAjaxError(e);
     }
 
-    this.dismissed = succeeded;
     this.disableControls = false;
   }
 
   @action
   async dismiss() {
     this.dismissed = true;
-    try {
-      await ajax("/newsletter-integration/subscriptions", { type: "DELETE" });
-    } catch (e) {
-      this.dismissed = false;
-      popupAjaxError(e);
+    if (!this.subscribed) {
+      try {
+        await ajax("/newsletter-integration/subscriptions", { type: "DELETE" });
+      } catch (e) {
+        this.dismissed = false;
+        popupAjaxError(e);
+      }
     }
   }
 }
